mvc/src/main/java/com/interface21/webmvc/servlet/mvc/HandlerAdapterRegistry.java
package com.interface21.webmvc.servlet.mvc;

import com.interface21.webmvc.servlet.mvc.exception.NotSupportedHandlerAdapterException;
import java.util.ArrayList;
import java.util.List;

public class HandlerAdapterRegistry {

    private final List<HandlerAdapter> handlerAdapters = new ArrayList<>();

    public void addHandlerAdapter(final HandlerAdapter handlerAdapter) {
        handlerAdapters.add(handlerAdapter);
    }

    public HandlerAdapter getHandlerAdapter(final Object handler) {
        return handlerAdapters.stream()
                .filter(handlerAdapter -> handlerAdapter.supports(handler))
                .findFirst()
                .orElseThrow(() -> new NotSupportedHandlerAdapterException(handler));
    }
}
